Fix undefined behaviour in od_worker_pool_feed function
Currently od_worker_pool_feed function uses non-atomic access to
pool->round_robin variable. But od_worker_pool_feed function can be
called multiple time simultaneously from od_system_server, as odyssey
supports multiple listen addresses feauture, and no synchronization is
performed between concurrent listen coroutines. This entails data race, which
leads undefined behaviour in C/C++

+++ sources/system.c
@@ -500,7 +500,7 @@ static inline void od_system(void *arg)
 	/* start worker threads */
 	od_worker_pool_t *worker_pool = system->global->worker_pool;
 	rc = od_worker_pool_start(worker_pool, system->global,
-				  instance->config.workers);
+				  (uint32_t)instance->config.workers);
 	if (rc == -1)
 		return;
